+++ R/add_p.R
@@ -124,7 +124,7 @@ add_p <- function(x, test = NULL, pvalue_fun = NULL,
   }
 
   # checking that input is class tbl_summary
-  if (class(x) != "tbl_summary") stop("x must be class 'tbl_summary'", call. = FALSE)
+  if (!inherits(x, "tbl_summary")) stop("`x` must be class 'tbl_summary'", call. = FALSE)
   # checking that input x has a by var
   if (is.null(x$inputs[["by"]])) {
     stop(paste0(